user level sync
(Logical change 1.1928)

+++ trace.h
@@ -18,8 +18,8 @@ typedef struct {
 	trace_file_type  type;
 	struct file     *fd;
 	char            *buf;
-	int              size;
-	int              used;
+	size_t           size;
+	size_t           used;
 	struct semaphore held;
 	int              disabled;
 } reiser4_trace_file;
@@ -32,14 +32,14 @@ typedef enum {
 } reiser4_traced_op;
 
 extern int open_trace_file   ( struct super_block *super,
-			       const char *file_name, int size, 
+			       const char *file_name, size_t size, 
 			       reiser4_trace_file *trace );
 extern int write_trace       ( reiser4_trace_file *file, 
 			       const char *format, ... ) 
 __attribute__ ((format (printf, 2, 3)));
 
 extern int write_trace_raw   ( reiser4_trace_file *file, 
-			       const void *data, int len );
+			       const void *data, size_t len );
 extern int hold_trace        ( reiser4_trace_file *file, int flag );
 extern int disable_trace     ( reiser4_trace_file *file, int flag );
 extern void close_trace_file ( reiser4_trace_file *file );
